requirements.txt
streamlit>=1.36
pandas>=2.2
python-dateutil>=2.9
beautifulsoup4>=4.12
requests>=2.32
ics>=0.7
pytz>=2024.1
jinja2>=3.1
openai>=1.35
playwright>=1.46
lxml>=5.2
reportlab>=4.1
